Add missiona address to ADOContract and move generation msg functions of Recipient to ado_base by creating a new trait
