+++ packages/ado_base/src/lib.rs
@@ -2,12 +2,13 @@ mod execute;
 #[cfg(feature = "instantiate")]
 mod instantiate;
 #[cfg(test)]
-pub mod mock_querier;
+mod mock_querier;
 #[cfg(feature = "modules")]
 pub mod modules;
 #[cfg(feature = "primitive")]
 pub mod primitive;
 mod query;
+pub mod recipient;
 pub mod state;
 #[cfg(feature = "withdraw")]
 mod withdraw;
