+++ packages/ado_base/src/state.rs
@@ -10,6 +10,7 @@ pub struct ADOContract<'a> {
     pub owner: Item<'a, Addr>,
     pub operators: Map<'a, &'a str, bool>,
     pub ado_type: Item<'a, String>,
+    pub mission_contract: Item<'a, Addr>,
     #[cfg(feature = "primitive")]
     pub primitive_contract: Item<'a, Addr>,
     #[cfg(feature = "modules")]
@@ -28,6 +29,7 @@ impl<'a> Default for ADOContract<'a> {
             owner: Item::new("owner"),
             operators: Map::new("operators"),
             ado_type: Item::new("ado_type"),
+            mission_contract: Item::new("mission_contract"),
             #[cfg(feature = "primitive")]
             primitive_contract: Item::new("primitive_contract"),
             #[cfg(feature = "modules")]
